Made the images clickable.
Changes to be committed:
modified:   src/components/Business/Business.js
In this file, the image is wrapped inside anchor tag with the
url pointing to the yelp site url which is taken from business
object. I also changed the variable associated with the address url from url to
addressurl. the website url is also deconstruucted and given to a
variable named websiteurl.

modified:   src/util/Yelp.js
In this file, the url element is returned inside the Buisness
object.