Image to links (#12)
* Added URL to Business element.

Added <a> tag to Business element which points to a google maps url to
open location of said element.

* Changed CSS of Link element in Business component

Set text decoration to none and text color to black to stick with the
style of the webpage

* removed unnecessary comments

* Made the images clickable.

Changes to be committed:
modified:   src/components/Business/Business.js
In this file, the image is wrapped inside anchor tag with the
url pointing to the yelp site url which is taken from business
object. I also changed the variable associated with the address url from url to
addressurl. the website url is also deconstruucted and given to a
variable named websiteurl.

modified:   src/util/Yelp.js
In this file, the url element is returned inside the Buisness
object.

* changed variable names and removed log fn